fix: try to delete nonexistent photo
if no photo is sent, an error is thrown and it's catched, but no photo is available to be deleted after catching: the deletion on error must only occur when a file is sent.

+++ api/src/controllers/photo.js
@@ -22,7 +22,7 @@ const store = async (req, res, next) => {
     const log = new Log(201, "Photo created", trace);
     res.json(photo);
   } catch (err) {
-    fs.unlink(`./api/uploads/images/${filename}`, () => {});
+    req.file && fs.unlink(`./api/uploads/images/${filename}`, () => {});
     const trace = stacktrace.parse(err);
     const errorContext = new ErrorContext(err, trace);
